Added more sources for in memory index, removed in memory iterators in favour of source iterators, tests pass

+++ core/src/main/java/org/lemurproject/galago/core/retrieval/iterator/disk/DiskLengthsIterator.java
@@ -46,7 +46,7 @@ public class DiskLengthsIterator extends SourceIterator
 
   @Override
   public int length() {
-    return (int) lengthSrc.length(context.document);
+    return lengthSrc.length(context.document);
   }
 
   @Override
